Alter `date.h` `hh_mm_ss` class by adding undocumented setters
Adds setters for `h_`, `m_`, and `s_.subseconds()`. I don't want to do this, but it is required by the custom parser to be able to set these fields individually. Collecting a new Duration object and using the constructor would normalize the individual components, which we don't want when parsing. The `s_.seconds()` setter was already exposed, so this isn't too much of a leap beyond that.