Allow `by` to be a duration
`by` is allowed to be a duration of equal precision or less precise precision than the common type of `from` and `to`. If it is a less precise precision, it also must be castable to the common type (i.e. can't use a `by` of 'year' precision when using 'day' precision input)