Update to latest tzdb (what will become tzdb 0.5.0) (#387)
* Apply `not_a_century` change https://github.com/HowardHinnant/date/commit/0ef86f1ced8788cf06f712e4ffbcb1cd74dacf65

* Apply `read_signed()` related patches

The upstream `date::read_signed()` has changed and now includes a `+` or `-` in the actual count of parsed values. This actually means that `%Y` can no longer parse `-1000` as a year without it explicitly being `%5Y`, which is something we need to add tests for if this change is here to stay.

Without this change, many tests break because `%z` no longer parses correctly. We preemptively look for `-` and `+`, but previously the fallthrough `read()` call would consume them without reducing the `M` max count and now it no longer does, so we have to `is.get()` to advance manually.

https://github.com/HowardHinnant/date/commit/ab37c362e35267d6dee02cb47760f9e9c669d3be

https://github.com/HowardHinnant/date/commit/5d225951ff24e6802d5026cb85d5b5a95b1a2476

* Apply dead code removal change https://github.com/HowardHinnant/date/commit/cd3c57932f7f3906c5876b8f04e2ff6b20d2070b

* Add remote on tzdb

* Let cpp11.R regenerate itself

* Apply `rs{}` to `ru{}` related changes https://github.com/HowardHinnant/date/commit/32ecb2ad24699963f6c58d74d74bcd8bf7e830a9

* NEWS bullet

* Fix indentation issue clangd identified

* Add test for negative year behavior

* NEWS bullet